// src/components/addressForm/Interface.ts
export interface AddressProps {
    id: number;
    uuid: string;
    title: string;
    zipcode: string;
    street_name: string;
    number: string;
    complement: string;
    neighbor: string;
    city: string;
    city_id: number;
    state: string;
    state_id: number;
    state_slug: string;
    lat: string;
    lng: string;
  }
  